chore: refactor actions_linux.go / execStarter
The execStarter function, which creates an EngineConfig defining how
to run a container, and then executes the starter binary, is extremely
long and complex. It accesses a lot of CLI flags directly, as global
variables, and is generally hard to reason about, especially where it
modifies those global vars in a way that changes behavior later.

The code shouldn't really be in the CLI layer. To handle execution of
OCI images with an OCI native runtime we need to be able to separate
the CLI from starter execution.

As a first step in refactoring to move this code out, this PR:

* Renames actions_linux.go to starter_linux.go to reflect the fact
  this code is all about calling the starter, and is not just for
  actions... it's used by instances too.
* Does some extraction of code into functions, to make execStarter
  itself easier to follow.
* Re-orders some code (where it is not order sensitive), to group like
  operations together (e.g. mount handling).

Note that the starter execution is still tied to the global vars
backing CLI flags, and modifies them. This is a first step of many :-)
